Move test suites according to the tested classes and delete redundant package.
